% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cities.R
\name{get_cities}
\alias{get_cities}
\title{Get cities around the world}
\usage{
get_cities(country_name = NULL, continent_name = NULL)
}
\arguments{
\item{country_name}{Optional, filter cities by country(ies).}

\item{continent_name}{Optional, filter cities by continent(s).}
}
\value{
a \code{data.table}
}
\description{
Get cities around the world
}
\examples{

all_cities <- get_cities()

asia_cities <- get_cities(continent_name = "Asia")

}
