% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cities.R
\name{get_continents}
\alias{get_continents}
\title{Get a summary of the continents of the cities}
\usage{
get_continents()
}
\value{
a \code{data.table}
}
\description{
Get a summary of the continents of the cities
}
\examples{

list_continents <- get_continents()

}
