% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cities.R
\name{get_countries}
\alias{get_countries}
\title{Get a summary of the countries of the cities}
\usage{
get_countries(continent_name = NULL)
}
\arguments{
\item{continent_name}{Optional, filter cities by continent(s).}
}
\value{
a \code{data.table}
}
\description{
Get a summary of the countries of the cities
}
\examples{

list_countries <- get_countries()

}
