% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{get_new_name}
\alias{get_new_name}
\title{Create a new string}
\usage{
get_new_name(x)
}
\arguments{
\item{x}{character vector}
}
\value{
a string not in `x`
}
\description{
Create a new string
}
\examples{
  get_new_name(letters)
}
