% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{get_wrapped_value}
\alias{get_wrapped_value}
\title{Extract wrapped value}
\usage{
get_wrapped_value(expr)
}
\arguments{
\item{expr}{an expression}
}
\value{
character if `is_wrapped(expr)`, otherwise `NULL`
}
\description{
Extract wrapped value
}
\examples{
  get_wrapped_value(substitute(a))
  get_wrapped_value(substitute({{a}}))
}
