% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{\%with\%}
\alias{\%with\%}
\title{Posterior variable declaration}
\usage{
expr \%with\% variables
}
\arguments{
\item{expr}{expression to evaluate}

\item{variables}{expression with variable assignments}
}
\value{
The value of the evaluated expression.
}
\description{
Posterior variable declaration
}
\examples{
  (a + b) \%with\% {
    a = 1
    b = 2
  }
}
