% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{is_wrapped}
\alias{is_wrapped}
\title{Check whether expression is 'wrapped' in "{{}}"}
\usage{
is_wrapped(expr)
}
\arguments{
\item{expr}{qan expression}
}
\value{
Boolean
}
\description{
Check whether expression is 'wrapped' in "{{}}"
}
\examples{
  is_wrapped(substitute(a))
  is_wrapped(substitute({{a}}))
}
