% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{make_substitutions}
\alias{make_substitutions}
\title{Process interpolated values}
\usage{
make_substitutions(expr)
}
\arguments{
\item{expr}{an expression}
}
\value{
list with (i) an expression, (ii) named character vector with names
  the interpolated variables, and values the new variable names.
}
\description{
Replaces values wrapped in `{{}}` with new variables.
}
