% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/where.R
\name{where}
\alias{where}
\title{Where is this City?}
\usage{
where(area = getOption("where.area"), time = getOption("where.playtime",
  default = 60))
}
\arguments{
\item{area}{Default selected country/continent.}

\item{time}{Playing time, default to 60 seconds.}
}
\description{
An addin to test your geographical skills !
}
\examples{
\dontrun{

where()

}
}
