% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/where.R
\name{where}
\alias{where}
\title{Where is this City?}
\usage{
where(mode = getOption("where.mode", default = "city"),
  area = getOption("where.area"), time = getOption("where.playtime",
  default = 60), viewer = getOption(x = "where.viewer", default =
  "dialog"))
}
\arguments{
\item{mode}{Play mode: \code{city} or \code{country}.}

\item{area}{Default selected country/continent.}

\item{time}{Playing time, default to 60 seconds.}

\item{viewer}{Where to display the gadget: \code{"dialog"},
\code{"pane"} or \code{"browser"} (see \code{\link[shiny]{viewer}}).}
}
\description{
An addin to test your geographical skills !
}
\examples{
\dontrun{

where()

}
}
